// src/interfaces/PagosSuspendidos.ts
export class PagosSuspendidos{
    Num_Poliza: string;
    Num_Orden: number
    Num_Iden: string;
    Nombre_Completo: string;
    FechaFallecimientoTitular : string;
    BanderaEnvio: number;
    BanderaCheck: number;
    identificador:number;
    FechaPagoReal: string;
    Meses:number;
    Monto:number;
  }